# Add resistance and impedance as weights to graph