--- a/pandapower/topology/create_graph.py
+++ b/pandapower/topology/create_graph.py
@@ -138,7 +138,7 @@ def create_nxgraph(net, respect_switches=True, include_lines=True, include_trafo
     else:
         # add edges for any bus-bus switches
         bs = net.switch[net.switch.et == "b"]
-    mg.add_edges_from((int(b), int(e), {"weight": 0, "key": int(i), "type": "s"})
+    mg.add_edges_from((int(b), int(e), {"weight": 0, "key": int(i), "type": "s", 'R': 0, 'Z': 0})
                       for b, e, i in list(zip(bs.bus, bs.element, bs.index)))
 
     # nogobuses are a nogo