For some analysis it is helpfull to have a measure of the electrical distance between buses and not only the length distance. Especially for networks with lots of impedances having only a weight of zero is not helpfull.

Be aware that I do not use the values directly so I am not 100 percent shure if my calculations are correct, please have a good look at the formulas I used. It gives reasonable results for the functionality I need though.