@lthurner 2018-07-05
This looks like sensible addition. Maybe we should add a parameter though that controls if weights like length, R, Z are added or not to avoid unecessary overhead, as I suspect in the great majority of cases this is not used. @ascheidl?