@WinfriedL 2018-07-05
Do you see memory or performance problems?

I wanted to test how much slower the addition would make the code. But strangely I got a faster graph creation.

I tried to create the graph 100 times for my testgrid with >6000 buses and nearly 10000 branches. The original implementation without r and impedance took ~24 seconds. The addition of resistance and impedance as weight took 14 seconds.

The only explanation I have at the moment is the change in how the three winding transformers are iterated...