@lthurner 2018-07-09
> Do you see memory or performance problems?

Since graph creation is one of those functions that might be carried out repeatedly a lot of times, so it should be optimized for performance. For example in our heuristic grid optimization, it is sometimes called several hundred of thousand times in one optimization, so performance is really important. I haven't explicitely checked the impact of your changes on performance though, just a note to keep performance in mind.

> I tried to create the graph 100 times for my testgrid with >6000 buses and nearly 10000 branches. The original implementation without r and impedance took ~24 seconds. The addition of resistance and impedance as weight took 14 seconds.

Your version of creating the 3W-transformers without the nested iterations is definitley better, so it makes sense it would be faster after the changes. But 14 seconds is still an extremely long time. I just tried the case6495rte network from the pandapower network package and the graph creation takes only 54ms... The network has >6000 nodes and >8500 branches (no 3W-transformers and no impedances though). If it really takes 14 seconds to create your grid graph, there has to be some serious bottleneck somewhere, probably still with the 3W-transformers?