@WinfriedL 2018-07-09
> The only explanation I have at the moment is the change in how the three winding transformers are iterated...

That is indeed the explanation. I did some more tests with creating the graph 100 times.

- Original implementation with iterrows for three winding transformers:
26 to 27 seconds

- Changing iterrows to itertuples: 
about 10.8 seconds

- Changing iterrows to explicitly create edges for all windings, similar to my pull request:
10.1 seconds

- Adding Resistance and Impedance (my pull request):
14 to 15 seconds.

So there is indeed an impact on performance by my addition, but only if you get rid of the iterrows in the first place  ;-)