@WinfriedL 2018-07-09
Did not see your reply before posting, sorry..

>  But do you really mean 14 seconds not milli seconds?

14 seconds for 100 times, so about 140 milli seconds per graph creation. That might be reasonable on my laptop...