@lthurner 2018-07-09
> 14 seconds for 100 times, so about 140 milli seconds per graph creation. That might be reasonable on my laptop...

Ok that makes more sense. But by your numbers adding the impedance as an edge weight increases runtime by 40-50% compared to the optimized version without the weights, which is definitely significant. So I would still argue for introducing a parameter in thre create_nxgraph function to have this as an optional feature