@WinfriedL 2018-07-09
If have added two parameters (default to false) for calculating r and z. If they are set to False, their weights are set to zero. It is not perfect but needed minimal changes...

I think there is still room for improvement, but for now it is faster than the original implementation even with both parameters set to True. And with both set to False I get between 100 and 110 ms for one graph creation.