@WinfriedL 2018-07-09
BTW, why are there the calls to list() in the creation? It seems unecessary to me...