@lthurner 2018-07-09
Faster and better, perfect!

> BTW, why are there the calls to list() in the creation? It seems unecessary to me...

If I remember correctly that is a Python 2 compatibility issue