@WinfriedL 2018-07-09
> If I remember correctly that is a Python 2 compatibility issue

At first sight it looks like it, but the compatibility problems are the other way around. In python 2 zip() returned a list, in python 3 an iterator. So list(zip()) in python3 is equivalent to zip() in python2, but since it is iterated over, it does not make any difference either way...